backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
django-storages==1.14.6
pillow==11.3.0
PyJWT==2.10.1
boto3==1.41.5
botocore==1.41.5
asgiref==3.10.0
sqlparse==0.5.3
python-dotenv==1.2.1